+++ apps/admin/src/app/_components/index.ts
@@ -1,0 +1,1 @@
+export { Builder } from './Builder'
